+++ src/index.ts
@@ -1,9 +1,11 @@
 import search from "./search";
 import random from "./random";
+import findById from "./find-by-id";
 import { resetCache } from "./spellbook-api";
 
 export = {
   search,
   random,
+  findById,
   resetCache,
 };